chore: scaffold core folder
