+++ src/core/add-gtag/index.ts
@@ -1,0 +1,1 @@
+export * from "./add-gtag";
