+++ src/core/settings/index.ts
@@ -1,0 +1,2 @@
+export * from "./settings";
+export * from "./types";
